Split the graphql file into two files for mutations and queries
