+++ frontend/src/components/MeetingView/MeetingView.jsx
@@ -4,7 +4,7 @@ import './MeetingView.scss';
 import { useParams } from 'react-router-dom';
 
 import { useLazyQuery, useQuery } from '@apollo/client';
-import { GET_MEETING_WITH_ITEMS } from '../../graphql/graphql';
+import { GET_MEETING_WITH_ITEMS } from '../../graphql/query';
 
 import NavBarHeader from '../NavBarHeader/NavBarHeader';
 import Header from '../Header/Header';
