+++ frontend/src/components/UserAccountCreate/AccountCreate.jsx
@@ -3,7 +3,8 @@ import './AccountCreate.styles.scss';
 import { useLazyQuery, useMutation } from '@apollo/client';
 import { Redirect } from 'react-router-dom';
 import ReCAPTCHA from 'react-google-recaptcha';
-import { GET_ACCOUNT_BY_EMAIL, LOGIN_LOCAL, CREATE_ACCOUNT } from '../../graphql/graphql';
+import { LOGIN_LOCAL, CREATE_ACCOUNT } from '../../graphql/mutation';
+import { GET_ACCOUNT_BY_EMAIL } from '../../graphql/query';
 import LocalStorageTerms from '../../constants/LocalStorageTerms';
 import LoginContext from '../LoginContext/LoginContext';
 import UserRoles from '../../constants/UserRoles';
